Implemented basic "security"

+++ src/main/java/controller/AdminController.java
@@ -9,9 +9,10 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
-import repository.book.BookRepository;
 import service.book.BookService;
 import service.user.UserService;
+
+import javax.servlet.http.HttpSession;
 
 @Controller
 public class AdminController {
@@ -24,41 +25,59 @@ public class AdminController {
 
     @RequestMapping(value = "/admin", method = RequestMethod.GET)
     @Order(value = 1)
-    public String index()
+    public String index(HttpSession session)
     {
+        if(!isLogged(session)){
+            return "redirect:/";
+        }
         return "admin";
     }
 
     @RequestMapping(value = "/admin/logout", method = RequestMethod.GET)
     @Order(value = 1)
-    public String logout()
+    public String logout(HttpSession session)
     {
+        if(!isLogged(session)){
+            return "redirect:/";
+        }
         return "redirect:/";
     }
 
     @RequestMapping(value = "/admin/users", method = RequestMethod.GET)
     @Order(value = 1)
-    public String usersIndex()
+    public String usersIndex(HttpSession session)
     {
+        if(!isLogged(session)){
+            return "redirect:/";
+        }
         return "manage-users";
     }
 
     @RequestMapping(value = "/admin/books", method = RequestMethod.GET)
     @Order(value = 1)
-    public String booksIndex()
+    public String booksIndex(HttpSession session)
     {
+        if(!isLogged(session)){
+            return "redirect:/";
+        }
         return "admin";
     }
 
     @RequestMapping(value = "/admin/users", method = RequestMethod.GET, params = {"action"})
-    public String viewAll(Model model) {
+    public String viewAll(Model model, HttpSession session) {
+        if(!isLogged(session)){
+            return "redirect:/";
+        }
         Iterable<User> users = userService.getAll();
         model.addAttribute("viewAll",users);
         return "manage-users";
     }
 
     @RequestMapping(value = "/admin/users", method = RequestMethod.POST, params = "action=create")
-    public String create(@RequestParam String uname, @RequestParam String upass, @RequestParam boolean uadmin, Model model){
+    public String create(@RequestParam String uname, @RequestParam String upass, @RequestParam boolean uadmin, Model model, HttpSession session){
+        if(!isLogged(session)){
+            return "redirect:/";
+        }
         Notification<Boolean> notification = userService.addUser(uname,upass,uadmin?1:0);
         if(notification.hasErrors()){
             model.addAttribute("result",notification.getFormattedErrors());
@@ -71,7 +90,10 @@ public class AdminController {
 
     @RequestMapping(value = "/admin/users", method = RequestMethod.POST, params = "action=update")
     public String update(@RequestParam int uid, @RequestParam String uname, @RequestParam String upass,
-                         @RequestParam boolean uadmin, Model model){
+                         @RequestParam boolean uadmin, Model model, HttpSession session){
+        if(!isLogged(session)){
+            return "redirect:/";
+        }
         Notification<Boolean> notification = userService.updateUser(uid,uname,upass,uadmin?1:0);
         if(notification.hasErrors()){
             model.addAttribute("result",notification.getFormattedErrors());
@@ -83,7 +105,10 @@ public class AdminController {
     }
 
     @RequestMapping(value = "/admin/users", method = RequestMethod.POST, params = "action=delete")
-    public String delete(@RequestParam int uid, Model model){
+    public String delete(@RequestParam int uid, Model model, HttpSession session){
+        if(!isLogged(session)){
+            return "redirect:/";
+        }
         Notification<Boolean> notification = userService.delete(uid);
         if(notification.hasErrors()){
             model.addAttribute("result",notification.getFormattedErrors());
@@ -92,5 +117,10 @@ public class AdminController {
         }
 
         return "manage-users";
+    }
+
+    private boolean isLogged(HttpSession session){
+        User user = (User) session.getAttribute("user");
+        return user != null && user.getIsAdmin() == 1;
     }
 }
